Login works with Retrofit call
Registration works with Retrofit call to Login/register/mobile

The Registration page is laid out with everything in it, the only things that are hardcoded now is the information coming back from the spinners.

There is validation for each of the fields before the call is made.

If the call successfully goes through, but is rejected by Spring for whatever reason, it will parse the reponse body, pull out the message and display it.
Eg. "This Email Address is already in use".

If the call and Registration succeeds it displays "Registration Successful" and an Intent brings you back to the "Login" page.